refactor: split points + query handlers into module dirs, dedup helpers
Split points.rs (533 lines) into points/mod.rs + points/streaming.rs.
Split query.rs (750 lines) into query/mod.rs + query/explain.rs +
query/velesql_helpers.rs. Extract error_response helper into helpers.rs.
Consolidate search pipeline helpers (parse_filter_or_400,
build_search_response) and reduce duplication across all handlers.

Net: -1325 lines across 11 server handler files.
