--- crates/velesdb-server/src/handlers/query.rs
@@ -1,750 +1,0 @@
-//! VelesQL query execution handler.
-
-use axum::{extract::State, http::StatusCode, response::IntoResponse, Json};
-use std::sync::Arc;
-
-use crate::types::{
-    AggregationResponse, ExplainCost, ExplainFeatures, ExplainRequest, ExplainResponse,
-    ExplainStep, QueryErrorDetail, QueryErrorResponse, QueryRequest, QueryResponse,
-    QueryResponseMeta, QueryType, SearchResultResponse, VelesqlErrorDetail, VelesqlErrorResponse,
-    VELESQL_CONTRACT_VERSION,
-};
-use crate::AppState;
-use velesdb_core::velesql::{self, Condition, Query, SelectColumns};
-
-fn is_aggregation_query(select: &velesdb_core::velesql::SelectStatement) -> bool {
-    matches!(
-        &select.columns,
-        SelectColumns::Aggregations(_) | SelectColumns::Mixed { .. }
-    ) || select.group_by.is_some()
-}
-
-fn aggregation_result_count(result: &serde_json::Value) -> usize {
-    match result {
-        serde_json::Value::Array(rows) => rows.len(),
-        serde_json::Value::Object(_) => 1,
-        _ => 0,
-    }
-}
-
-#[allow(deprecated)]
-fn execute_aggregation_query(
-    state: &Arc<AppState>,
-    collection_name: &str,
-    parsed: &Query,
-    params: &std::collections::HashMap<String, serde_json::Value>,
-    start: std::time::Instant,
-) -> axum::response::Response {
-    let collection = match state.db.get_collection(collection_name) {
-        Some(c) => c,
-        None => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(VelesqlErrorResponse {
-                    error: VelesqlErrorDetail {
-                        code: "VELESQL_COLLECTION_NOT_FOUND".to_string(),
-                        message: format!("Collection '{}' not found", collection_name),
-                        hint: "Create the collection first or correct the collection name"
-                            .to_string(),
-                        details: Some(serde_json::json!({
-                            "collection": collection_name
-                        })),
-                    },
-                }),
-            )
-                .into_response()
-        }
-    };
-
-    let result = match collection.execute_aggregate(parsed, params) {
-        Ok(r) => r,
-        Err(e) => {
-            return (
-                StatusCode::UNPROCESSABLE_ENTITY,
-                Json(VelesqlErrorResponse {
-                    error: VelesqlErrorDetail {
-                        code: "VELESQL_AGGREGATION_ERROR".to_string(),
-                        message: e.to_string(),
-                        hint: "Verify GROUP BY/HAVING clauses and aggregate function arguments"
-                            .to_string(),
-                        details: None,
-                    },
-                }),
-            )
-                .into_response()
-        }
-    };
-
-    let timing_ms = start.elapsed().as_secs_f64() * 1000.0;
-    let duration_us = start.elapsed().as_micros();
-    #[allow(clippy::cast_possible_truncation)]
-    state.db.notify_query(
-        collection_name,
-        duration_us.min(u128::from(u64::MAX)) as u64,
-    );
-    let count = aggregation_result_count(&result);
-
-    Json(AggregationResponse {
-        result,
-        timing_ms,
-        meta: QueryResponseMeta {
-            velesql_contract_version: VELESQL_CONTRACT_VERSION.to_string(),
-            count,
-        },
-    })
-    .into_response()
-}
-
-/// Execute a VelesQL query.
-///
-/// BUG-1 FIX: Automatically detects aggregation queries (GROUP BY, COUNT, SUM, etc.)
-/// and routes them to execute_aggregate for proper handling.
-#[utoipa::path(
-    post,
-    path = "/query",
-    tag = "query",
-    request_body = QueryRequest,
-    responses(
-        (status = 200, description = "Query results", body = QueryResponse),
-        (status = 400, description = "Query syntax error", body = QueryErrorResponse),
-        (status = 422, description = "Query validation/execution error", body = VelesqlErrorResponse),
-        (status = 404, description = "Collection not found", body = VelesqlErrorResponse)
-    )
-)]
-#[allow(clippy::unused_async, deprecated)]
-pub async fn query(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<QueryRequest>,
-) -> impl IntoResponse {
-    let start = std::time::Instant::now();
-
-    let parsed = match parse_and_validate(&req.query) {
-        Ok(q) => q,
-        Err(resp) => return resp,
-    };
-
-    let collection_name = match resolve_collection_name(&parsed, &req) {
-        Ok(name) => name,
-        Err(resp) => return resp,
-    };
-
-    // BUG-1 FIX: Detect aggregation queries and route to execute_aggregate
-    if is_aggregation_query(&parsed.select) {
-        return execute_aggregation_query(&state, &collection_name, &parsed, &req.params, start);
-    }
-
-    let results = match execute_standard_query(&state, &parsed, &collection_name, &req) {
-        Ok(r) => r,
-        Err(resp) => return resp,
-    };
-
-    build_query_response(&state, &collection_name, start, results)
-}
-
-/// Parse a VelesQL query string and run validation, returning an error response on failure.
-#[allow(clippy::result_large_err)]
-fn parse_and_validate(query_str: &str) -> Result<Query, axum::response::Response> {
-    let parsed = velesql::Parser::parse(query_str).map_err(|e| {
-        (
-            StatusCode::BAD_REQUEST,
-            Json(QueryErrorResponse {
-                error: QueryErrorDetail {
-                    error_type: format!("{:?}", e.kind),
-                    message: e.message.clone(),
-                    position: e.position,
-                    query: e.fragment.clone(),
-                },
-            }),
-        )
-            .into_response()
-    })?;
-
-    velesql::QueryValidator::validate(&parsed).map_err(|e| {
-        (
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(VelesqlErrorResponse {
-                error: VelesqlErrorDetail {
-                    code: "VELESQL_VALIDATION_ERROR".to_string(),
-                    message: e.to_string(),
-                    hint: e.suggestion,
-                    details: None,
-                },
-            }),
-        )
-            .into_response()
-    })?;
-
-    Ok(parsed)
-}
-
-/// Determine the target collection from the parsed query and request body.
-#[allow(clippy::result_large_err)]
-fn resolve_collection_name(
-    parsed: &Query,
-    req: &QueryRequest,
-) -> Result<String, axum::response::Response> {
-    if parsed.is_match_query() {
-        req.collection
-            .as_ref()
-            .filter(|name| !name.is_empty())
-            .cloned()
-            .ok_or_else(|| {
-                (
-                    StatusCode::UNPROCESSABLE_ENTITY,
-                    Json(VelesqlErrorResponse {
-                        error: VelesqlErrorDetail {
-                            code: "VELESQL_MISSING_COLLECTION".to_string(),
-                            message: "MATCH query via /query requires `collection` in request body"
-                                .to_string(),
-                            hint: "Add `collection` to the /query JSON body or use /collections/{name}/match".to_string(),
-                            details: Some(serde_json::json!({
-                                "field": "collection",
-                                "endpoint": "/query",
-                                "query_type": "MATCH"
-                            })),
-                        },
-                    }),
-                )
-                    .into_response()
-            })
-    } else {
-        Ok(parsed.select.from.clone())
-    }
-}
-
-/// Execute a standard (non-aggregation) query, dispatching MATCH vs SELECT.
-#[allow(deprecated, clippy::result_large_err)]
-fn execute_standard_query(
-    state: &Arc<AppState>,
-    parsed: &Query,
-    collection_name: &str,
-    req: &QueryRequest,
-) -> Result<Vec<velesdb_core::SearchResult>, axum::response::Response> {
-    let execute_result = if parsed.is_match_query() {
-        match state.db.get_collection(collection_name) {
-            Some(c) => c.execute_query(parsed, &req.params),
-            None => Err(velesdb_core::Error::CollectionNotFound(
-                collection_name.to_string(),
-            )),
-        }
-    } else {
-        state.db.execute_query(parsed, &req.params)
-    };
-
-    execute_result.map_err(|e| match e {
-        velesdb_core::Error::CollectionNotFound(name) => (
-            StatusCode::NOT_FOUND,
-            Json(VelesqlErrorResponse {
-                error: VelesqlErrorDetail {
-                    code: "VELESQL_COLLECTION_NOT_FOUND".to_string(),
-                    message: format!("Collection '{}' not found", name),
-                    hint: "Create the collection first or correct the collection name".to_string(),
-                    details: Some(serde_json::json!({ "collection": name })),
-                },
-            }),
-        )
-            .into_response(),
-        other => (
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(VelesqlErrorResponse {
-                error: VelesqlErrorDetail {
-                    code: "VELESQL_EXECUTION_ERROR".to_string(),
-                    message: other.to_string(),
-                    hint:
-                        "Validate query semantics and parameter types against the target collection"
-                            .to_string(),
-                    details: None,
-                },
-            }),
-        )
-            .into_response(),
-    })
-}
-
-/// Build the final query response with timing metrics.
-fn build_query_response(
-    state: &Arc<AppState>,
-    collection_name: &str,
-    start: std::time::Instant,
-    results: Vec<velesdb_core::SearchResult>,
-) -> axum::response::Response {
-    let timing_ms = start.elapsed().as_secs_f64() * 1000.0;
-    #[allow(clippy::cast_possible_truncation)]
-    let took_ms = timing_ms.round() as u64;
-    let duration_us = start.elapsed().as_micros();
-    #[allow(clippy::cast_possible_truncation)]
-    state.db.notify_query(
-        collection_name,
-        duration_us.min(u128::from(u64::MAX)) as u64,
-    );
-    let rows_returned = results.len();
-
-    Json(QueryResponse {
-        results: results
-            .into_iter()
-            .map(|r| SearchResultResponse {
-                id: r.point.id,
-                score: r.score,
-                payload: r.point.payload,
-            })
-            .collect(),
-        timing_ms,
-        took_ms,
-        rows_returned,
-        meta: QueryResponseMeta {
-            velesql_contract_version: VELESQL_CONTRACT_VERSION.to_string(),
-            count: rows_returned,
-        },
-    })
-    .into_response()
-}
-
-/// Execute an aggregation-only VelesQL query.
-///
-/// This endpoint is explicit and stable for GROUP BY / HAVING / aggregate workloads.
-#[utoipa::path(
-    post,
-    path = "/aggregate",
-    tag = "query",
-    request_body = QueryRequest,
-    responses(
-        (status = 200, description = "Aggregation results", body = AggregationResponse),
-        (status = 400, description = "Query syntax error", body = QueryErrorResponse),
-        (status = 422, description = "Aggregation validation/execution error", body = VelesqlErrorResponse),
-        (status = 404, description = "Collection not found", body = VelesqlErrorResponse)
-    )
-)]
-#[allow(clippy::unused_async)]
-pub async fn aggregate(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<QueryRequest>,
-) -> impl IntoResponse {
-    let start = std::time::Instant::now();
-
-    let parsed = match velesql::Parser::parse(&req.query) {
-        Ok(q) => q,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(QueryErrorResponse {
-                    error: QueryErrorDetail {
-                        error_type: format!("{:?}", e.kind),
-                        message: e.message.clone(),
-                        position: e.position,
-                        query: e.fragment.clone(),
-                    },
-                }),
-            )
-                .into_response()
-        }
-    };
-
-    if let Err(e) = velesql::QueryValidator::validate(&parsed) {
-        return (
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(VelesqlErrorResponse {
-                error: VelesqlErrorDetail {
-                    code: "VELESQL_VALIDATION_ERROR".to_string(),
-                    message: e.to_string(),
-                    hint: e.suggestion,
-                    details: None,
-                },
-            }),
-        )
-            .into_response();
-    }
-
-    if parsed.is_match_query() || !is_aggregation_query(&parsed.select) {
-        return (
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(VelesqlErrorResponse {
-                error: VelesqlErrorDetail {
-                    code: "VELESQL_AGGREGATION_ERROR".to_string(),
-                    message: "Only aggregation queries are accepted on /aggregate".to_string(),
-                    hint: "Use /query for row/search/graph queries; use /aggregate for GROUP BY/aggregate workloads.".to_string(),
-                    details: Some(serde_json::json!({
-                        "endpoint": "/aggregate"
-                    })),
-                },
-            }),
-        )
-            .into_response();
-    }
-
-    let collection_name = if parsed.select.from.is_empty() {
-        match req.collection.as_ref().filter(|name| !name.is_empty()) {
-            Some(name) => name.clone(),
-            None => {
-                return (
-                    StatusCode::UNPROCESSABLE_ENTITY,
-                    Json(VelesqlErrorResponse {
-                        error: VelesqlErrorDetail {
-                            code: "VELESQL_MISSING_COLLECTION".to_string(),
-                            message: "Aggregation query requires a FROM collection or request-body `collection`".to_string(),
-                            hint: "Add FROM <collection> to query or set `collection` in request JSON".to_string(),
-                            details: Some(serde_json::json!({
-                                "field": "collection",
-                                "endpoint": "/aggregate"
-                            })),
-                        },
-                    }),
-                )
-                    .into_response()
-            }
-        }
-    } else {
-        parsed.select.from.clone()
-    };
-
-    execute_aggregation_query(&state, &collection_name, &parsed, &req.params, start)
-}
-
-/// Explain a VelesQL query without executing it (EPIC-058 US-002).
-///
-/// Returns the query plan, estimated costs, and detected features.
-#[utoipa::path(
-    post,
-    path = "/query/explain",
-    tag = "query",
-    request_body = ExplainRequest,
-    responses(
-        (status = 200, description = "Query plan", body = ExplainResponse),
-        (status = 400, description = "Query syntax error", body = QueryErrorResponse),
-        (status = 404, description = "Collection not found", body = VelesqlErrorResponse)
-    )
-)]
-#[allow(clippy::unused_async, deprecated)]
-pub async fn explain(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<ExplainRequest>,
-) -> impl IntoResponse {
-    let parsed = match velesql::Parser::parse(&req.query) {
-        Ok(q) => q,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(QueryErrorResponse {
-                    error: QueryErrorDetail {
-                        error_type: format!("{:?}", e.kind),
-                        message: e.message.clone(),
-                        position: e.position,
-                        query: e.fragment.clone(),
-                    },
-                }),
-            )
-                .into_response()
-        }
-    };
-
-    let select = &parsed.select;
-
-    let collection_exists = state.db.get_collection(&select.from).is_some();
-    if !collection_exists && !select.from.is_empty() {
-        return (
-            StatusCode::NOT_FOUND,
-            Json(VelesqlErrorResponse {
-                error: VelesqlErrorDetail {
-                    code: "VELESQL_COLLECTION_NOT_FOUND".to_string(),
-                    message: format!("Collection '{}' not found", select.from),
-                    hint: "Create the collection first or correct the FROM collection".to_string(),
-                    details: Some(serde_json::json!({ "collection": select.from })),
-                },
-            }),
-        )
-            .into_response();
-    }
-
-    let features = detect_explain_features(select);
-    let plan = build_explain_plan(select, &features);
-    let estimated_cost = estimate_cost(features.has_vector_search);
-
-    let query_type = if parsed.is_match_query() {
-        "MATCH"
-    } else {
-        "SELECT"
-    };
-
-    let (cache_hit, plan_reuse_count) = state
-        .db
-        .explain_query(&parsed)
-        .ok()
-        .map_or((None, None), |qp| (qp.cache_hit, qp.plan_reuse_count));
-
-    Json(ExplainResponse {
-        query: req.query,
-        query_type: query_type.to_string(),
-        collection: select.from.clone(),
-        plan,
-        estimated_cost,
-        features,
-        cache_hit,
-        plan_reuse_count,
-    })
-    .into_response()
-}
-
-/// Detect query features from a SELECT statement for EXPLAIN output.
-fn detect_explain_features(select: &velesdb_core::velesql::SelectStatement) -> ExplainFeatures {
-    let has_vector_search = select
-        .where_clause
-        .as_ref()
-        .map(condition_has_vector_search)
-        .unwrap_or(false);
-
-    ExplainFeatures {
-        has_vector_search,
-        has_filter: select.where_clause.is_some() && !has_vector_search,
-        has_order_by: select.order_by.is_some(),
-        has_group_by: select.group_by.is_some(),
-        has_aggregation: matches!(
-            &select.columns,
-            SelectColumns::Aggregations(_) | SelectColumns::Mixed { .. }
-        ),
-        has_join: !select.joins.is_empty(),
-        has_fusion: select.fusion_clause.is_some(),
-        limit: select.limit,
-        offset: select.offset,
-    }
-}
-
-/// Build the execution plan steps for an EXPLAIN response.
-fn build_explain_plan(
-    select: &velesdb_core::velesql::SelectStatement,
-    features: &ExplainFeatures,
-) -> Vec<ExplainStep> {
-    let mut plan = Vec::new();
-    let mut step_num = 1;
-
-    plan.push(build_source_step(select, features, step_num));
-    step_num += 1;
-
-    append_filter_and_join_steps(select, features, &mut plan, &mut step_num);
-    append_aggregation_steps(features, &mut plan, &mut step_num);
-    append_pagination_step(select, &mut plan, step_num);
-
-    plan
-}
-
-fn build_source_step(
-    select: &velesdb_core::velesql::SelectStatement,
-    features: &ExplainFeatures,
-    step_num: usize,
-) -> ExplainStep {
-    if features.has_vector_search {
-        ExplainStep {
-            step: step_num,
-            operation: "VectorSearch".to_string(),
-            description: "ANN search using HNSW index with NEAR clause".to_string(),
-            estimated_rows: select.limit.map(|l| l as usize),
-        }
-    } else {
-        ExplainStep {
-            step: step_num,
-            operation: "FullScan".to_string(),
-            description: format!("Scan collection '{}'", select.from),
-            estimated_rows: None,
-        }
-    }
-}
-
-fn append_filter_and_join_steps(
-    select: &velesdb_core::velesql::SelectStatement,
-    features: &ExplainFeatures,
-    plan: &mut Vec<ExplainStep>,
-    step_num: &mut usize,
-) {
-    if features.has_filter {
-        plan.push(ExplainStep {
-            step: *step_num,
-            operation: "Filter".to_string(),
-            description: "Apply WHERE clause predicates".to_string(),
-            estimated_rows: None,
-        });
-        *step_num += 1;
-    }
-
-    for join in &select.joins {
-        plan.push(ExplainStep {
-            step: *step_num,
-            operation: format!("{:?}Join", join.join_type),
-            description: format!("Join with '{}'", join.table),
-            estimated_rows: None,
-        });
-        *step_num += 1;
-    }
-}
-
-fn append_aggregation_steps(
-    features: &ExplainFeatures,
-    plan: &mut Vec<ExplainStep>,
-    step_num: &mut usize,
-) {
-    if features.has_group_by {
-        plan.push(ExplainStep {
-            step: *step_num,
-            operation: "GroupBy".to_string(),
-            description: "Group rows by specified columns".to_string(),
-            estimated_rows: None,
-        });
-        *step_num += 1;
-    }
-
-    if features.has_aggregation {
-        plan.push(ExplainStep {
-            step: *step_num,
-            operation: "Aggregate".to_string(),
-            description: "Compute aggregate functions (COUNT, SUM, etc.)".to_string(),
-            estimated_rows: None,
-        });
-        *step_num += 1;
-    }
-
-    if features.has_order_by {
-        plan.push(ExplainStep {
-            step: *step_num,
-            operation: "Sort".to_string(),
-            description: "Sort results by ORDER BY clause".to_string(),
-            estimated_rows: None,
-        });
-        *step_num += 1;
-    }
-}
-
-fn append_pagination_step(
-    select: &velesdb_core::velesql::SelectStatement,
-    plan: &mut Vec<ExplainStep>,
-    step_num: usize,
-) {
-    if select.limit.is_some() || select.offset.is_some() {
-        plan.push(ExplainStep {
-            step: step_num,
-            operation: "Limit".to_string(),
-            description: format!(
-                "Apply LIMIT {} OFFSET {}",
-                select.limit.unwrap_or(0),
-                select.offset.unwrap_or(0)
-            ),
-            estimated_rows: select.limit.map(|l| l as usize),
-        });
-    }
-}
-
-/// Estimate execution cost based on query features.
-fn estimate_cost(has_vector_search: bool) -> ExplainCost {
-    ExplainCost {
-        uses_index: has_vector_search,
-        index_name: if has_vector_search {
-            Some("HNSW".to_string())
-        } else {
-            None
-        },
-        selectivity: if has_vector_search { 0.01 } else { 1.0 },
-        complexity: if has_vector_search {
-            "O(log n)"
-        } else {
-            "O(n)"
-        }
-        .to_string(),
-    }
-}
-
-/// Check if a condition contains vector search.
-fn condition_has_vector_search(cond: &Condition) -> bool {
-    match cond {
-        Condition::VectorSearch(_)
-        | Condition::VectorFusedSearch { .. }
-        | Condition::Similarity(_) => true,
-        Condition::And(left, right) | Condition::Or(left, right) => {
-            condition_has_vector_search(left) || condition_has_vector_search(right)
-        }
-        Condition::Group(inner) | Condition::Not(inner) => condition_has_vector_search(inner),
-        _ => false,
-    }
-}
-
-/// Detect query type from parsed AST (EPIC-052 US-006).
-///
-/// Priority order:
-/// 1. MATCH clause → Graph
-/// 2. GROUP BY or aggregates → Aggregation
-/// 3. Vector search → Search
-/// 4. Default → Rows
-#[allow(dead_code)] // Used in tests, will be used in unified handler
-pub fn detect_query_type(query: &Query) -> QueryType {
-    // Check for MATCH clause first
-    if query.is_match_query() {
-        return QueryType::Graph;
-    }
-
-    let select = &query.select;
-
-    // Check for aggregation
-    let is_aggregation = matches!(
-        &select.columns,
-        SelectColumns::Aggregations(_) | SelectColumns::Mixed { .. }
-    ) || select.group_by.is_some();
-
-    if is_aggregation {
-        return QueryType::Aggregation;
-    }
-
-    // Check for vector search
-    let has_vector = select
-        .where_clause
-        .as_ref()
-        .map(condition_has_vector_search)
-        .unwrap_or(false);
-
-    if has_vector {
-        return QueryType::Search;
-    }
-
-    QueryType::Rows
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_detect_query_type_search() {
-        let parsed = velesql::Parser::parse(
-            "SELECT * FROM docs WHERE similarity(embedding, $v) > 0.8 LIMIT 10",
-        )
-        .unwrap();
-        assert_eq!(detect_query_type(&parsed), QueryType::Search);
-    }
-
-    #[test]
-    fn test_detect_query_type_aggregation() {
-        let parsed =
-            velesql::Parser::parse("SELECT category, COUNT(*) FROM products GROUP BY category")
-                .unwrap();
-        assert_eq!(detect_query_type(&parsed), QueryType::Aggregation);
-    }
-
-    #[test]
-    fn test_detect_query_type_rows() {
-        let parsed =
-            velesql::Parser::parse("SELECT name, price FROM products WHERE price > 100").unwrap();
-        assert_eq!(detect_query_type(&parsed), QueryType::Rows);
-    }
-
-    #[test]
-    fn test_detect_query_type_graph() {
-        let parsed =
-            velesql::Parser::parse("MATCH (n:Person)-[:KNOWS]->(m) RETURN n.name, m.name LIMIT 10")
-                .unwrap();
-        assert_eq!(detect_query_type(&parsed), QueryType::Graph);
-    }
-
-    #[test]
-    fn test_detect_query_type_hybrid_vector_aggregation() {
-        // When both vector search and aggregation, aggregation takes priority
-        let parsed = velesql::Parser::parse(
-            "SELECT category, COUNT(*) FROM docs WHERE similarity(embedding, $v) > 0.7 GROUP BY category",
-        )
-        .unwrap();
-        assert_eq!(detect_query_type(&parsed), QueryType::Aggregation);
-    }
-}
